# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GSB Ticket Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/tickets:
    post:
      operationId: createAdminTicket
      tags: [admin]
      summary: Crear un registro manualmente (solo administración)
      description: >
        Alta manual de un registro desde el panel de administración. El flujo
        normal de creación de tickets sigue siendo el webhook de n8n.
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Registro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "401":
          description: Clave de administración inválida
        "409":
          description: Ya existe un ticket con ese conversation_id
  /admin/import:
    post:
      operationId: importCsv
      tags: [admin]
      summary: Importación masiva desde CSV
      description: >
        Recibe el contenido de un CSV (texto plano), detecta las columnas con
        los mismos alias que el importador CLI, y crea los registros que no
        existan. Idempotente por conversation_id. Con dry_run true solo simula
        y devuelve el resumen sin escribir nada.
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminImportInput"
      responses:
        "200":
          description: Resumen de la importación (o de la simulación)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminImportResult"
        "400":
          description: CSV inválido o sin columna conversation_id
        "401":
          description: Clave de administración inválida
  /admin/truncate:
    post:
      operationId: truncateTickets
      tags: [admin]
      summary: Borrar todos los registros (truncate)
      description: >
        Elimina TODOS los tickets y seguimientos y reinicia los contadores de
        id. El schema queda intacto. Requiere confirmar explícitamente.
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTruncateInput"
      responses:
        "200":
          description: Registros eliminados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTruncateResult"
        "400":
          description: Falta la confirmación explícita
        "401":
          description: Clave de administración inválida
  /admin/roles:
    get:
      operationId: listAdminRoles
      tags: [admin]
      summary: Listar roles
      security:
        - adminApiKey: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Lista paginada de roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRoleListResponse"
        "400":
          description: Parametros de consulta invalidos
        "401":
          description: Clave de administracion invalida

    post:
      operationId: createAdminRole
      tags: [admin]
      summary: Crear un rol
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRoleInput"
      responses:
        "201":
          description: Rol creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRole"
        "400":
          description: Cuerpo invalido
        "401":
          description: Clave de administracion invalida
        "409":
          description: Ya existe un rol con ese nombre
  /admin/roles/{id}:
    patch:
      operationId: updateAdminRole
      tags: [admin]
      summary: Actualizar un rol
      security:
        - adminApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRoleUpdate"
      responses:
        "200":
          description: Rol actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRole"
        "400":
          description: Id o cuerpo invalido
        "401":
          description: Clave de administracion invalida
        "404":
          description: Rol no encontrado
        "409":
          description: Ya existe un rol con ese nombre

    delete:
      operationId: deleteAdminRole
      tags: [admin]
      summary: Eliminar un rol sin usuarios asignados
      security:
        - adminApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "204":
          description: Rol eliminado
        "400":
          description: Id invalido
        "401":
          description: Clave de administracion invalida
        "404":
          description: Rol no encontrado
        "409":
          description: El rol tiene usuarios asignados
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: Listar usuarios
      security:
        - adminApiKey: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: role_id
          in: query
          schema:
            type: integer
            minimum: 1
        - name: activo
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Lista paginada de usuarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "400":
          description: Parametros de consulta invalidos
        "401":
          description: Clave de administracion invalida

    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Crear un usuario
      security:
        - adminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Cuerpo invalido o rol inexistente
        "401":
          description: Clave de administracion invalida
        "409":
          description: Ya existe un usuario con ese email
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Actualizar o desactivar un usuario
      description: Los usuarios se desactivan estableciendo activo en false; no se eliminan.
      security:
        - adminApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Id o cuerpo invalido
        "401":
          description: Clave de administracion invalida
        "404":
          description: Usuario o rol no encontrado
        "409":
          description: Ya existe un usuario con ese email
...
components:
  securitySchemes:
    webhookApiKey:
      type: apiKey
      in: header
      name: x-api-key
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
    AdminRole:
      type: object
      required:
        - id
        - nombre
        - descripcion
        - activo
        - fecha_creacion
        - fecha_actualizacion
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        activo:
          type: boolean
        fecha_creacion:
          type: string
          format: date-time
        fecha_actualizacion:
          type: string
          format: date-time
    AdminRoleInput:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        descripcion:
          type: ["string", "null"]
          maxLength: 500
        activo:
          type: boolean
          default: true
    AdminRoleUpdate:
      type: object
      minProperties: 1
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        descripcion:
          type: ["string", "null"]
          maxLength: 500
        activo:
          type: boolean
    AdminRoleListResponse:
      type: object
      required:
        - roles
        - total
        - page
        - limit
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/AdminRole"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUser:
      type: object
      required:
        - id
        - nombre
        - apellido
        - email
        - role_id
        - activo
        - fecha_creacion
        - fecha_actualizacion
      properties:
        id:
          type: integer
        nombre:
          type: string
        apellido:
          type: ["string", "null"]
        email:
          type: string
        role_id:
          type: integer
        activo:
          type: boolean
        fecha_creacion:
          type: string
          format: date-time
        fecha_actualizacion:
          type: string
          format: date-time
    AdminUserInput:
      type: object
      required:
        - nombre
        - email
        - role_id
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        apellido:
          type: ["string", "null"]
          maxLength: 100
        email:
          type: string
          pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'
          maxLength: 254
        role_id:
          type: integer
          minimum: 1
        activo:
          type: boolean
          default: true
    AdminUserUpdate:
      type: object
      minProperties: 1
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        apellido:
          type: ["string", "null"]
          maxLength: 100
        email:
          type: string
          pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$'
          maxLength: 254
        role_id:
          type: integer
          minimum: 1
        activo:
          type: boolean
    AdminUserListResponse:
      type: object
      required:
        - users
        - total
        - page
        - limit
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminImportInput:
      type: object
      required:
        - csv
      properties:
        csv:
          type: string
          description: Contenido completo del archivo CSV (texto plano)
        dry_run:
          type: boolean
          default: false
          description: Si es true, simula la importación sin escribir nada
    AdminImportColumna:
      type: object
      required:
        - columna
        - campo
      properties:
        columna:
          type: string
          description: Encabezado original del CSV
        campo:
          type: string
          description: Campo del ticket al que se mapeó
    AdminImportResult:
      type: object
      required:
        - dry_run
        - filas
        - insertados
        - ya_existentes
        - invalidos
        - columnas
        - sin_mapear
        - advertencias
      properties:
        dry_run:
          type: boolean
        filas:
          type: integer
          description: Filas de datos leídas del CSV
        insertados:
          type: integer
        ya_existentes:
          type: integer
        invalidos:
          type: integer
        columnas:
          type: array
          items:
            $ref: "#/components/schemas/AdminImportColumna"
        sin_mapear:
          type: array
          items:
            type: string
        advertencias:
          type: array
          items:
            type: string
    AdminTruncateInput:
      type: object
      required:
        - confirmar
      properties:
        confirmar:
          type: boolean
          description: Debe ser true para ejecutar el borrado
    AdminTruncateResult:
      type: object
      required:
        - tickets_eliminados
        - seguimientos_eliminados
      properties:
        tickets_eliminados:
          type: integer
        seguimientos_eliminados:
          type: integer
...
    Ticket:
      type: object
      required:
        - id
        - conversation_id
        - hora
        - nombre
        - apellido
        - estado
        - prioridad
        - motivo
        - fecha_creacion
        - notificado
      properties:
        id:
          type: integer
        conversation_id:
          type: string
        hora:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: ["string", "null"]
        dni:
          type: ["string", "null"]
        empresa:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        motivo:
          type: string
        resumen:
          type: ["string", "null"]
        notificado:
          type: boolean
        estado:
          type: string
          enum: [nuevo, en_proceso, pendiente, resuelto, cerrado]
        prioridad:
          type: string
          enum: [baja, media, alta, urgente]
        asignado_a:
          type: ["string", "null"]
        audio_url:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
        fecha_creacion:
          type: string
          format: date-time
        fecha_limite:
          type: ["string", "null"]
          format: date-time
        fecha_resolucion:
          type: ["string", "null"]
          format: date-time
        progreso:
          type: integer
          minimum: 0
          maximum: 100
...
    TicketInput:
      type: object
      required:
        - conversation_id
        - hora
        - nombre
        - apellido
        - motivo
      properties:
        conversation_id:
          type: string
        hora:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
        dni:
          type: string
        empresa:
          type: string
        email:
          type: string
        motivo:
          type: string
        resumen:
          type: string
        notificado:
          type: boolean
          default: false
        estado:
          type: string
          enum: [nuevo, en_proceso, pendiente, resuelto, cerrado]
          default: nuevo
        prioridad:
          type: string
          enum: [baja, media, alta, urgente]
          default: media
        asignado_a:
          type: string
        audio_url:
          type: string
        notas:
          type: string
        fecha_limite:
          type: string
          format: date-time
        progreso:
          type: integer
          minimum: 0
          maximum: 100
          default: 0
...