# lib/api-zod/src/generated/types/adminRole.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GSB Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRole {
  id: number;
  nombre: string;
  /** @nullable */
  descripcion: string | null;
  activo: boolean;
  fecha_creacion: Date;
  fecha_actualizacion: Date;
}