# lib/api-zod/src/generated/types/adminRoleListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GSB Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface AdminRoleListResponse {
  roles: AdminRole[];
  total: number;
  page: number;
  limit: number;
}