# lib/api-zod/src/generated/types/adminRoleUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GSB Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminRoleUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  nombre?: string;
  /**
     * @maxLength 500
     * @nullable
     */
  descripcion?: string | null;
  activo?: boolean;
}