# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * GSB Ticket Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  nombre: string;
  /** @nullable */
  apellido: string | null;
  email: string;
  role_id: number;
  activo: boolean;
  fecha_creacion: Date;
  fecha_actualizacion: Date;
}